clickhouse: upgrade to clickhouse-go/v2

+++ middleware/mysql/global_test.go
@@ -49,7 +49,7 @@ func TestGlobalPool(t *testing.T) {
 	asst.Nil(err, "execute sql with global pool failed")
 	actual, err := result.(*Result).GetIntByName(0, "ok")
 	asst.Nil(err, "execute sql with global pool failed")
-	asst.Equal(int64(1), actual, "expected and actual values are not equal")
+	asst.Equal(1, actual, "expected and actual values are not equal")
 
 	// sleep to testStruct maintain mechanism
 	time.Sleep(10 * time.Second)
